+++ nailmirror/src/mock/eval-hands.js
@@ -3,79 +3,79 @@ module.exports = [
   {
     "id": "eval-1",
     "label": "评测手照1",
-    "handUrl": "http://p0.meituan.net/pilotimages/b9632e3a699fdb63a1a6139bbfd6bf0d2159483.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/87797733466cfd525625a5947767e2ff1794125.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/b9632e3a699fdb63a1a6139bbfd6bf0d2159483.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/87797733466cfd525625a5947767e2ff1794125.png"
   },
   {
     "id": "eval-2",
     "label": "评测手照2",
-    "handUrl": "http://p1.meituan.net/pilotimages/704b1c4bdf589b5d5367f2748f6868f42205269.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/162afb52255bd908ba3ec418fd61824a2254875.png"
+    "handUrl": "https://p1.meituan.net/pilotimages/704b1c4bdf589b5d5367f2748f6868f42205269.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/162afb52255bd908ba3ec418fd61824a2254875.png"
   },
   {
     "id": "eval-3",
     "label": "评测手照3",
-    "handUrl": "http://p0.meituan.net/pilotimages/3cd4bc446f321574df68ce0a749b16b62603765.png",
-    "pairStyleUrl": "http://p1.meituan.net/pilotimages/7bb5bc0c2c741f9f0aa63787a601d7ad2604877.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/3cd4bc446f321574df68ce0a749b16b62603765.png",
+    "pairStyleUrl": "https://p1.meituan.net/pilotimages/7bb5bc0c2c741f9f0aa63787a601d7ad2604877.png"
   },
   {
     "id": "eval-4",
     "label": "评测手照4",
-    "handUrl": "http://p0.meituan.net/pilotimages/7c791f4b4b13659d62d991f172f5ffd02674881.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/fc8fe60e78341d77a5070fc2f8e520072098070.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/7c791f4b4b13659d62d991f172f5ffd02674881.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/fc8fe60e78341d77a5070fc2f8e520072098070.png"
   },
   {
     "id": "eval-5",
     "label": "评测手照5",
-    "handUrl": "http://p1.meituan.net/pilotimages/5a7efacd78020469ab44e4caca1afe972676586.png",
-    "pairStyleUrl": "http://p1.meituan.net/pilotimages/3c0d090e20f0cb56f70fcb56c54dd6582416974.png"
+    "handUrl": "https://p1.meituan.net/pilotimages/5a7efacd78020469ab44e4caca1afe972676586.png",
+    "pairStyleUrl": "https://p1.meituan.net/pilotimages/3c0d090e20f0cb56f70fcb56c54dd6582416974.png"
   },
   {
     "id": "eval-6",
     "label": "评测手照6",
-    "handUrl": "http://p1.meituan.net/pilotimages/6a3d032df4a143c79c3e2ec3cd4c53522723999.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/6c857edd85a5fa4bcec59698fe9416cb1913981.png"
+    "handUrl": "https://p1.meituan.net/pilotimages/6a3d032df4a143c79c3e2ec3cd4c53522723999.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/6c857edd85a5fa4bcec59698fe9416cb1913981.png"
   },
   {
     "id": "eval-7",
     "label": "评测手照7",
-    "handUrl": "http://p0.meituan.net/pilotimages/ed9a1cd3cca3997ede3779771dda6a772149757.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/2ac2d01a9bc78320edbe2b545b485b4a2132292.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/ed9a1cd3cca3997ede3779771dda6a772149757.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/2ac2d01a9bc78320edbe2b545b485b4a2132292.png"
   },
   {
     "id": "eval-8",
     "label": "评测手照8",
-    "handUrl": "http://p1.meituan.net/pilotimages/a52c995f1f9e2e668c6093099cfd24032514125.png",
-    "pairStyleUrl": "http://p1.meituan.net/pilotimages/d15c06e8c2137d4f39f3b60476a90cf92026957.png"
+    "handUrl": "https://p1.meituan.net/pilotimages/a52c995f1f9e2e668c6093099cfd24032514125.png",
+    "pairStyleUrl": "https://p1.meituan.net/pilotimages/d15c06e8c2137d4f39f3b60476a90cf92026957.png"
   },
   {
     "id": "eval-9",
     "label": "评测手照9",
-    "handUrl": "http://p0.meituan.net/pilotimages/e0d094b37e9595e465280b26b9cefc7b2318139.png",
-    "pairStyleUrl": "http://p1.meituan.net/pilotimages/69614397f0ecb559b98cb46a5a46f3b32642714.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/e0d094b37e9595e465280b26b9cefc7b2318139.png",
+    "pairStyleUrl": "https://p1.meituan.net/pilotimages/69614397f0ecb559b98cb46a5a46f3b32642714.png"
   },
   {
     "id": "eval-10",
     "label": "评测手照10",
-    "handUrl": "http://p0.meituan.net/pilotimages/4b310e7ce87af3d5f20064a25420f34b2320446.png",
-    "pairStyleUrl": "http://p1.meituan.net/pilotimages/2277d6f9d82264fa6a3c986373e5e44c2292083.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/4b310e7ce87af3d5f20064a25420f34b2320446.png",
+    "pairStyleUrl": "https://p1.meituan.net/pilotimages/2277d6f9d82264fa6a3c986373e5e44c2292083.png"
   },
   {
     "id": "eval-11",
     "label": "评测手照11",
-    "handUrl": "http://p0.meituan.net/pilotimages/e763917e1e5b4a68f33d8075fc1504bf2680556.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/bc153edf655dd6961dc9f8e95ad8cd1e2561531.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/e763917e1e5b4a68f33d8075fc1504bf2680556.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/bc153edf655dd6961dc9f8e95ad8cd1e2561531.png"
   },
   {
     "id": "eval-12",
     "label": "评测手照12",
-    "handUrl": "http://p0.meituan.net/pilotimages/8a73302ae05d6520b90d50db36d481492489372.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/43cc4ced977a3dd271f60ee2f05607772681747.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/8a73302ae05d6520b90d50db36d481492489372.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/43cc4ced977a3dd271f60ee2f05607772681747.png"
   },
   {
     "id": "eval-13",
     "label": "评测手照13",
-    "handUrl": "http://p0.meituan.net/pilotimages/badf638f6e3989a31d524b7a8cc4a2332706287.png",
-    "pairStyleUrl": "http://p0.meituan.net/pilotimages/682c173ae3a95d0b838655e8337b30d72213857.png"
+    "handUrl": "https://p0.meituan.net/pilotimages/badf638f6e3989a31d524b7a8cc4a2332706287.png",
+    "pairStyleUrl": "https://p0.meituan.net/pilotimages/682c173ae3a95d0b838655e8337b30d72213857.png"
   }
 ];